refactor : option portal 컴포넌트 위치 수정 및 적용
