+++ src/components/OptionPortal/HashTag.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useRecoilState } from 'recoil';
 import { v4 as getRandomKey } from 'uuid';
 
-import * as S from '@components/ShareForm/Option/OptionPortal/OptionPortal.style';
+import * as S from '@components/OptionPortal/OptionPortal.style';
 import Icon from '@components/common/Icon';
 import { tagList } from '@store/shareRegistration';
